+++ Lesson12/pom.xml
@@ -2,13 +2,14 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>c18onlTMS2022</artifactId>
         <groupId>by.tms</groupId>
+        <artifactId>c18onlTMS2022</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
-    <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>Lesson5</artifactId>
+    <groupId>by.tms</groupId>
+    <artifactId>Lesson12</artifactId>
 
 </project>
