+++ Lesson12/src/main/java/by/tms/Main.java
@@ -1,0 +1,7 @@
+package by.tms;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
