+++ Lesson3/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>c18onlTMS2022</artifactId>
-        <groupId>org.example</groupId>
+        <groupId>by.tms</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
