+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
+    <groupId>by.tms</groupId>
     <artifactId>c18onlTMS2022</artifactId>
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
@@ -24,6 +24,7 @@
         <module>Lesson9Inheritance</module>
         <module>Lesson10</module>
         <module>Lesson11</module>
+        <module>Lesson12</module>
     </modules>
 
     <properties>
